# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Твой API Справочник Организаций",
        "description": "API для работы с организациями, зданиями и видами деятельности",
        "contact": {
            "email": "твой@email.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер"
        }
    ],
    "paths": {
...
        "/api/organizations/by-name": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций по названию",
                "description": "Находит организации, название которых содержит указанную строку (минимум 3 символа). Возвращает список организаций с телефонами и данными здания. Если ничего не найдено, возвращается пустой список.",
                "operationId": "getOrganizationsByName",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Часть названия организации (минимум 3 символа).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 3
                        },
                        "example": "Рем"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Список организаций (может быть пустым)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationResource"
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Название должно содержать минимум 3 символа."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Название должно содержать минимум 3 символа."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/organizations/nearby": {
            "get": {
                "tags": [
                    "Организации"
                ],
                "summary": "Поиск организаций в заданном радиусе",
                "description": "Находит организации, расположенные в указанном радиусе от заданной географической точки. Возвращает список организаций с телефонными номерами.",
                "operationId": "getOrganizationsInRadius",
                "parameters": [
                    {
                        "name": "api_key",
                        "in": "query",
                        "description": "Ключ API для аутентификации.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "123"
                    },
                    {
                        "name": "lat",
                        "in": "query",
                        "description": "Широта центральной точки (от -90 до 90).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 90,
                            "minimum": -90
                        },
                        "example": 55.7069
                    },
                    {
                        "name": "lng",
                        "in": "query",
                        "description": "Долгота центральной точки (от -180 до 180).",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "maximum": 180,
                            "minimum": -180
                        },
                        "example": 37.5422026
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Радиус поиска в метрах (минимум 0).",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 0
                        },
                        "example": 1000
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успех. Список организаций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationWithPhonesResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Организации не найдены"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The lat field is required."
                                        },
                                        "errors": {
                                            "properties": {
                                                "lat": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The lat field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "OrganizationPhoneResource": {
                "title": "Ресурс телефонного номера организации",
                "description": "Представление контактного телефонного номера организации",
                "required": [
                    "id",
                    "number"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор",
                        "description": "Уникальный числовой идентификатор телефонного номера",
                        "type": "integer",
                        "example": 1
                    },
                    "number": {
                        "title": "Номер телефона",
                        "description": "Контактный телефонный номер в международном или национальном формате",
                        "type": "string",
                        "example": "+7 (495) 123-45-67"
                    }
                },
                "type": "object"
            },
            "OrganizationResource": {
                "title": "Базовый ресурс организации",
                "description": "Минимальный набор данных для идентификации организации",
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "title": "Идентификатор организации",
                        "description": "Уникальный числовой идентификатор организации в системе",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "title": "Наименование",
                        "description": "Официальное название организации",
                        "type": "string",
                        "example": "ООО 'Промышленные технологии'"
                    }
                },
                "type": "object"
            },
            "OrganizationWithPhonesResource": {
                "title": "Ресурс организации с телефонами",
                "description": "Данные организации, включая идентификатор, наименование и список телефонных номеров.",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/OrganizationResource"
                    },
                    {
                        "properties": {
                            "phones": {
                                "title": "Телефоны",
                                "description": "Список телефонных номеров, связанных с организацией.",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/OrganizationPhoneResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "api_key",
                "in": "query"
            }
        }
    },
...
}